refactor notificationCard
